requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
youtube-transcript-api
google-generativeai
python-dotenv
